fixed forced shell and bash bug in Makefile and 32bit compilation

+++ src/display_test_result.c
@@ -191,8 +191,8 @@ static char	*get_speed_color(_Bool inverted) {
 		char	*speeder = (vdiff > 1) ? "slower" : "faster";
 		//if (vdiff < 1)
 		//	vdiff = 1 / vdiff;
-		dprintf(g_log_fd, "%s: x%.2f %s than %s's function (%lu cpu ticks vs %lu)\n",
-				current_fun_name, vdiff, speeder, opponent, TIME_DIFF_LIB, TIME_DIFF_SYS);
+		dprintf(g_log_fd, "%s: x%.2f %s than %s's function (%llu cpu ticks vs %llu)\n",
+				current_fun_name, vdiff, speeder, opponent, (unsigned long long)TIME_DIFF_LIB, (unsigned long long)TIME_DIFF_SYS);
 	}
 	if (diff > 10)
 		return (COLOR_SPEED_10);
